Feat mesh textures (#145)
* Created dynamic buffers by resizing

* Fixed memory read bounds and formatting

* Fixed comments on PR

* Formatted

* Meshes working with textures

* Use normals by default on GLTF load and correct directional light usage

* Formatted files

+++ src/engine_core/rendering/src/WebGPU/WebGPUGraphicsDevice.cpp
@@ -277,7 +277,7 @@ namespace Hush::Graphics
 		HUSH_ASSERT(
 			Bitwise::HasCompositeFlag(newDesc.usage, EBufferUsage::CopySource),
 			"For a buffer to be resized it needs to have the CopySource usage flag, current buffer's usage flags: {}",
-			newDesc.usage);
+			static_cast<uint32_t>(newDesc.usage));
 		newDesc.size = size;
 
 		// Copy the original buffer back to the new one
